move flattenObject into util functions file
